Refactored Composables to Components.kt
